% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_computations.R
\name{add_U_H}
\alias{add_U_H}
\title{Appends a list of Ui and Hi matrices}
\usage{
add_U_H(matrix_list, bread, Ufun = wgee::U_standard)
}
\arguments{
\item{matrix_list}{}

\item{bread}{result of the \code{\link{bread}} function}
}
\description{
Appends a list of Ui and Hi matrices
}

