% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{bread}
\alias{bread}
\title{Bread maker}
\usage{
bread(matrix_list, geeglm_ingredients)
}
\arguments{
\item{matrix_list}{list of matrices}

\item{geeglm_ingredients}{result of \code{\link{extract_geeglm}}}
}
\description{
Computes
\deqn{ \sum_{i = 1}^N D_i^T V_i^{-1} W_i D_i}
}

