% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_geeglm}
\alias{extract_geeglm}
\title{Extract necessary elements from geeglm object}
\usage{
extract_geeglm(geeobj)
}
\arguments{
\item{geeobj}{}
}
\description{
Extract necessary elements from geeglm object
}

