% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geeglm_w.R
\name{geeglm_w}
\alias{geeglm_w}
\title{geeglm_w}
\usage{
geeglm_w(formula, data, id, family = gaussian, correction = "none", ...)
}
\arguments{
\item{geeobj}{}

\item{Ufun}{}
}
\description{
geeglm_w
}

