% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{Hi_matrix}
\alias{Hi_matrix}
\title{Cluster-level H matrix}
\usage{
Hi_matrix(D, V.inv, bread)
}
\arguments{
\item{D}{matrix of derivatives of mean function}

\item{V.inv}{inverse variance matrix}

\item{bread}{result of the \code{\link{bread}} function}
}
\description{
Computes
\deqn{ D_i O D_i^T V_^{-1}}
}
\details{
where \eqn{O} is the result of the \code{\link{bread}} function
}

