% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{make_cluster_matrices}
\alias{make_cluster_matrices}
\title{Make cluster level matrices necessary for computations}
\usage{
make_cluster_matrices(df)
}
\arguments{
\item{df}{the data.frame in \code{\link{extract_geeglm}}. It must contain
the following variables in the following order: \code{response}, \code{fitted}, \code{weights}.
Remaining (covariate) variables must begin at column index 4.}
}
\description{
Make cluster level matrices necessary for computations
}

