% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{O_matrix}
\alias{O_matrix}
\title{Cluster-level O matrix}
\usage{
O_matrix(D, W, V.inv)
}
\arguments{
\item{D}{matrix of derivatives of mean function}

\item{W}{weight matrix}

\item{V.inv}{inverse variance matrix}
}
\description{
Computes
\deqn{ D_i^T V_i^{-1} W_i D_i}
}

