% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_computations.R
\name{sqrt_matrix}
\alias{sqrt_matrix}
\title{Taken from: http://www4.stat.ncsu.edu/~li/software/SparseSDR.R
square-root-inverse of a matrix}
\usage{
sqrt_matrix(M)
}
\description{
Taken from: http://www4.stat.ncsu.edu/~li/software/SparseSDR.R
square-root-inverse of a matrix
}

