% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{summerizer}
\alias{summerizer}
\title{Sums a list of matrices}
\usage{
summerizer(matrix_list, geeglm_ingredients, solve = FALSE)
}
\arguments{
\item{matrix_list}{list of matrices to sum across (must all be of same dimensions)}

\item{geeglm_ingredients}{result of \code{\link{extract_geeglm}}}
}
\description{
Sums a list of matrices
}

