% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_computations.R
\name{U_MD}
\alias{U_MD}
\title{Cluster-level MD-corrected U matrix}
\usage{
U_MD(parts)
}
\arguments{
\item{parts}{}
}
\description{
Computes
\deqn{ D_i^T V_^{-1} (I_i - H_i)^{-1} W_i r_i}
}

