% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_computations.R
\name{U_standard}
\alias{U_standard}
\title{Cluster-level U matrix}
\usage{
U_standard(parts)
}
\arguments{
\item{D}{matrix of derivatives of mean function}

\item{V.inv}{inverse variance matrix}

\item{W}{matrix of weight}

\item{r}{matrix of residuals}
}
\description{
Computes
\deqn{ D_i^T V_^{-1} W_i r_i}
}

