% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_computations.R
\name{veggies}
\alias{veggies}
\title{Meat maker}
\usage{
veggies(matrix_list, geeglm_ingredients)
}
\arguments{
\item{matrix_list}{list of matrices}

\item{geeglm_ingredients}{result of \code{\link{extract_geeglm}}}
}
\description{
Computes
\deqn{ \sum_{i = 1}^N U_i U_i^T}
}

