% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{U_MD}
\alias{U_MD}
\title{Cluster-level KC-corrected U matrix}
\usage{
U_MD(parts)
}
\arguments{
\item{D}{matrix of derivatives of mean function}

\item{V.inv}{inverse variance matrix}

\item{W}{matrix of weight}

\item{r}{matrix of residuals}
}
\description{
Computes
\deqn{ D_i^T V_^{-1} W_i r_i}
}

